Chackra UI Setup

+++ 100-hours-web-development-bootcamp-build-23-react-projects/mern-crash-course/backend/controllers/product.controller.js
@@ -31,6 +31,10 @@ export const createProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isInvalid(id)) {
+    return res.status(404).json({ success: false, message: 'Invalid Product id' });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
